backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.134.0",
    "google-genai>=1.65.0",
    "pydantic-settings>=2.13.1",
]
